Adding defaultValue attribute to <Property>
This allows components with Properties to be used even if value is not
set on building, e.g. with <Assign> element. Useful for synapses with
weight Property, to avoid having 2 definitions, one with & one without
weight.

Adding menu item on GUI to dump current state type/instance info to
console